Add a check for the potential breaking changes warning
